requirements.txt
Django==1.10.4
djangorestframework==3.5.3
wheel==0.24.0
drfdocs